requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
Cython==3.0.12
Flask==3.1.0
gevent==24.11.1
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-speech==2.31.1
google-cloud-storage==3.1.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
requests==2.32.3
rsa==4.9
setuptools==77.0.3
# speex==0.9.1
urllib3==2.3.0
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
groq==0.25.0
dotenv==0.9.9